requirements.txt
google-generativeai
rich
prompt_toolkit
python-dotenv